// frontend/src/app/core/api/model/setPasswordHintRequest.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SetPasswordHintRequest { 
    masterPassword: string;
    hint?: string;
}