// frontend/src/app/core/api/model/updateSecurityQuestionsRequest.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SecurityQuestionDTO } from './securityQuestionDTO';


export interface UpdateSecurityQuestionsRequest { 
    masterPassword: string;
    securityQuestions: Array<SecurityQuestionDTO>;
}